+++ Runtime/Scripts/ScriptableEvent/Events/BoolEvent.cs
@@ -9,11 +9,11 @@ namespace Dev.Nicklaj.Butter
     {
         public bool HasArgument => true;
 
-        public void OnRaiseButtonSubmit(string arg0 = "")
+        public void OnRaiseButtonSubmit(string arg0 = "", uint channel = 0)
         {
             // Try to parse the input as a boolean
             if (bool.TryParse(arg0, out bool result) == false) return;
-            this.Raise(result);
+            this.Raise(result, channel);
         }
     }
 }